Add some api

+++ Controllers/CartItemsController.cs
@@ -1,8 +1,10 @@
 ﻿using BenriShop.ApiRepository.CartItems;
 using BenriShop.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace BenriShop.Controllers
@@ -22,10 +24,16 @@ namespace BenriShop.Controllers
         /// </summary>
         /// <param name="userName"></param>
         /// <returns></returns>
-        // GET: api/CartItems/userName
-        [HttpGet("{userName}")]
+        // GET: api/CartItems/GetCartItems/userName
+        [Authorize(Roles = "Customer")]
+        [HttpGet("GetCartItems/{userName}")]
         public async Task<IEnumerable<CartItem>> GetCartItems(string userName)
         {
+            var identity = User.Identity as ClaimsIdentity;
+            if (identity.Name != userName)
+            {
+                return (IEnumerable<CartItem>)Conflict("Can't access to diffirent account");
+            }
             var cartItems = _cartItemRepository.GetCartItems(userName);
             if (cartItems != null)
             {
@@ -40,12 +48,18 @@ namespace BenriShop.Controllers
         /// <param name="productId"></param>
         /// <param name="cartItem"></param>
         /// <returns></returns>
-        // PUT: api/CartItems/userName/1
+        // PUT: api/CartItems/UpdateCartItem/userName/1
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
-        [HttpPut("{userName}/{productId}")]
+        [Authorize(Roles = "Customer")]
+        [HttpPut("UpdateCartItem/{userName}/{productId}")]
         public async Task<IActionResult> UpdateCartItem(string userName, int productId, CartItem cartItem)
         {
+            var identity = User.Identity as ClaimsIdentity;
+            if (identity.Name != userName)
+            {
+                return Conflict("Can't access to diffirent account");
+            }
             if (productId != cartItem.ProductId || userName != cartItem.UserName)
             {
                 return BadRequest("Wrong of productId or userName");
@@ -69,16 +83,22 @@ namespace BenriShop.Controllers
         /// </summary>
         /// <param name="cartItem"></param>
         /// <returns></returns>
-        // POST: api/CartItems
+        // POST: api/CartItems/AddCartItem
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
-        [HttpPost]
+        [Authorize(Roles = "Customer")]
+        [HttpPost("AddCartItem")]
         public async Task<ActionResult<CartItem>> AddCartItem(CartItem cartItem)
         {
+            var identity = User.Identity as ClaimsIdentity;
+            if (identity.Name != cartItem.UserName)
+            {
+                return NotFound("Can't accesss to diffirent account");
+            }
             try
             {
                 await _cartItemRepository.AddCartItem(cartItem);
-                return CreatedAtAction("GetCartItems", new { userName = cartItem.UserName }, cartItem);
+                return Ok("Add cart item is successful");
             }catch
             {
                 return BadRequest("Error when call _cartItemRepository.AddCartItem(cartItem)");
@@ -91,11 +111,17 @@ namespace BenriShop.Controllers
         /// <param name="userName"></param>
         /// <param name="productId"></param>
         /// <returns></returns>
-        // DELETE: api/CartItems/5
-        [HttpDelete("{userName}/{productId}")]
+        // DELETE: api/CartItems/DeleteCartItem/userName/1
+        [Authorize(Roles = "Customer")]
+        [HttpDelete("DeleteCartItem/{userName}/{productId}")]
         public async Task<ActionResult<CartItem>> DeleteCartItem(string userName, int productId)
         {
-            if(await _cartItemRepository.DeleteCartItem(userName, productId))
+            var identity = User.Identity as ClaimsIdentity;
+            if (identity.Name != userName)
+            {
+                return NotFound("Can't accesss to diffirent account");
+            }
+            if (await _cartItemRepository.DeleteCartItem(userName, productId))
             {
                 return Ok("Delete successfully!");
             }
